+++ GenumKit/Stencil/GenumTemplate.swift
@@ -45,6 +45,7 @@ open class GenumTemplate: Template {
 func genumExtension() -> Extension {
   let namespace = Extension()
   namespace.registerTag("set", parser: SetNode.parse)
+  namespace.registerTag("map", parser: MapNode.parse)
   namespace.registerFilter("swiftIdentifier", filter: StringFilters.stringToSwiftIdentifier)
   namespace.registerFilter("join", filter: ArrayFilters.join)
   namespace.registerFilter("lowerFirstWord", filter: StringFilters.lowerFirstWord)